requirements.txt
mysql.connector
pandas
random